+++ packages/swc-plugin-workflow/transform/tests/fixture/class-expression-binding-name-step-methods/output-step.js
@@ -22,7 +22,7 @@ var LanguageModel = function(__wf_cls) {
     });
     var __wf_cls_sym = Symbol.for("workflow-class-registry"), __wf_cls_reg = globalThis[__wf_cls_sym] || (globalThis[__wf_cls_sym] = new Map());
     __wf_cls_reg.set("class//./input//LanguageModel", __wf_cls);
-    Object.defineProperty(__wf_cls, "classId", {
+    if (!Object.prototype.hasOwnProperty.call(__wf_cls, "classId")) Object.defineProperty(__wf_cls, "classId", {
         value: "class//./input//LanguageModel",
         writable: false,
         enumerable: false,